resources: Add X86 Spatter disk image
Adds packer script that builds a disk image that contains
Spatter + JSON input files that are representative of LANL
workloads.

Builds using QEMU and cloud-init; Proxy can be set in the
shared/user-data file and it will be propagated through the
packer build

Also builds m5 inside the packer script instead of copying
it in; Costs a little under 300MB of space, most of which
is the .git folder from the gem5 repo.